Add new components and integrate Redux
This commit includes several new components for the app including DetailedPostView, FilterOptions, PostItem, PostList, and SearchAppBar. Redux has also been integrated into the application along with the addition of the Roboto font. The sorting and filtering functionality as well as the search bar in the app header are yet to be implemented.