Update UX/UI interactions across components
This commit introduces multiple User Experience and User Interface changes across various components. Highlight changes include updating the AppBar to handle clicks on the Reddit Icon and transitioning it to a fixed position, updating the color toggling mechanism in PostItem, and adding a new theme in index.js. The Footer has been fully revamped to include information about the developer and links to social media. In FilterOptions, the selected subreddit is dynamically handled, and the component's position has been fixed in Home.